handle rename of everything
