+++ tests/web/rpc/v1/test_endpoint_delete_trace_items.py
@@ -3,14 +3,14 @@ import uuid
 import pytest
 from google.protobuf.json_format import MessageToDict
 from google.protobuf.timestamp_pb2 import Timestamp
-from sentry_protos.snuba.v1.endpoint_delete_trace_pb2 import (
-    DeleteTraceRequest,
-    DeleteTraceResponse,
+from sentry_protos.snuba.v1.endpoint_delete_trace_items_pb2 import (
+    DeleteTraceItemsRequest,
+    DeleteTraceItemsResponse,
 )
 from sentry_protos.snuba.v1.request_common_pb2 import RequestMeta, ResponseMeta
 
 from snuba.web.rpc.common.exceptions import BadSnubaRPCRequestException
-from snuba.web.rpc.v1.endpoint_delete_trace import EndpointDeleteTrace
+from snuba.web.rpc.v1.endpoint_delete_trace_items import EndpointDeleteTraceItems
 from tests.base import BaseApiTest
 
 _REQUEST_ID = uuid.uuid4().hex
@@ -22,7 +22,7 @@ class TestEndpointDeleteTrace(BaseApiTest):
     def test_missing_trace_id_raises_exception(self) -> None:
         ts = Timestamp()
         ts.GetCurrentTime()
-        message = DeleteTraceRequest(
+        message = DeleteTraceItemsRequest(
             meta=RequestMeta(
                 project_ids=[1, 2, 3],
                 organization_id=1,
@@ -36,7 +36,7 @@ class TestEndpointDeleteTrace(BaseApiTest):
         )
 
         with pytest.raises(BadSnubaRPCRequestException) as exc_info:
-            EndpointDeleteTrace().execute(message)
+            EndpointDeleteTraceItems().execute(message)
 
         assert "trace_id is required for deleting a trace." in str(exc_info.value)
 
@@ -44,7 +44,7 @@ class TestEndpointDeleteTrace(BaseApiTest):
         ts = Timestamp()
         ts.GetCurrentTime()
         trace_id = uuid.uuid4().hex
-        message = DeleteTraceRequest(
+        message = DeleteTraceItemsRequest(
             meta=RequestMeta(
                 project_ids=[1, 2, 3],
                 organization_id=1,
@@ -57,9 +57,9 @@ class TestEndpointDeleteTrace(BaseApiTest):
             trace_id=trace_id,
         )
 
-        response = EndpointDeleteTrace().execute(message)
+        response = EndpointDeleteTraceItems().execute(message)
 
-        expected_response = DeleteTraceResponse(
+        expected_response = DeleteTraceItemsResponse(
             meta=ResponseMeta(request_id=_REQUEST_ID),
             matching_items_count=0,
         )
